fix: move genesis helpers to app helpers
