+++ app/app.go
@@ -1,7 +1,6 @@
 package kiichain
 
 import (
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/fs"
@@ -64,8 +63,6 @@ import (
 	evmencoding "github.com/cosmos/evm/encoding"
 	srvflags "github.com/cosmos/evm/server/flags"
 	cosmosevmtypes "github.com/cosmos/evm/types"
-	erc20types "github.com/cosmos/evm/x/erc20/types"
-	evmtypes "github.com/cosmos/evm/x/vm/types"
 
 	kiiante "github.com/kiichain/kiichain/v5/ante"
 	"github.com/kiichain/kiichain/v5/app/keepers"
@@ -562,21 +559,6 @@ func (app *KiichainApp) GetBaseApp() *baseapp.BaseApp {
 // GetTxConfig implements the TestingApp interface.
 func (app *KiichainApp) GetTxConfig() client.TxConfig {
 	return app.txConfig
-}
-
-// DefaultGenesis returns a default genesis from the registered ModuleBasics's.
-func (app *KiichainApp) DefaultGenesis() map[string]json.RawMessage {
-	genesis := app.ModuleBasics.DefaultGenesis(app.appCodec)
-
-	evmGenState := NewEVMGenesisState()
-	genesis[evmtypes.ModuleName] = app.appCodec.MustMarshalJSON(evmGenState)
-
-	// NOTE: for the example chain implementation we are also adding a default token pair,
-	// which is the base denomination of the chain (i.e. the WEVMOS contract)
-	erc20GenState := NewErc20GenesisState()
-	genesis[erc20types.ModuleName] = app.appCodec.MustMarshalJSON(erc20GenState)
-
-	return genesis
 }
 
 // GetTestGovKeeper implements the TestingApp interface.
